+++ primitives/src/epoch.rs
@@ -15,10 +15,12 @@ pub enum EpochNumber {
     Number(u64),
     /// Earliest block (checkpoint).
     Earliest,
-    /// Latest mined block.
-    LatestMined,
+    /// The latest checkpoint (cur_era_genesis)
+    LatestCheckpoint,
     /// Latest block with state.
     LatestState,
+    /// Latest mined block.
+    LatestMined,
 }
 
 impl Into<EpochNumber> for u64 {