+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Improve the performance of the consensus layer for unstable TreeGraph scenarios. 
 
 
+- Change the default value of `from_epoch` in RPC `cfx_getLogs` from "earliest" to "latest_checkpoint".
+Now if no `from_epoch` is specified, it will only return logs after the latest checkpoint.
+
 # 0.5.0
 
 ## Improvements